Edit index.js file to include all code in one file; delete 2 JS files.
